requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord.py==1.7.1
feedparser==6.0.2
idna==2.10
multidict==5.1.0
oauthlib==3.1.0
praw==7.2.0
prawcore==2.0.0
pycodestyle==2.7.0
PySocks==1.7.1
python-dotenv==0.17.0
requests==2.25.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
toml==0.10.2
tweepy==3.10.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
yarl==1.6.3
